ISPN-10698 Remove InternalDataContainerAdapter
* Also remove AbstractDelegatingDataContainer
